need to review the existing function calls and signals/slots between hexwidget, reader, slider and main app to ensure there are no int transfers which will kill the new slider...
implement an up down arrow button as well for the slider.

+++ wombatforensics.h
@@ -741,7 +741,8 @@ private slots:
     {
         statuslabel->setText(tmptext);
     };
-    void SetStepValues(int singlestep, int pagestep);
+    //void SetStepValues(int singlestep, int pagestep);
+    void SetStepValues(off_t singlestep, off_t pagestep);
     void TreeContextMenu(const QPoint &point);
     void ImgHexMenu(const QPoint &point);
     void SetFilter(int headercolumn);
